+++ src/Tests/Dfe.Complete.CypressTests/cypress/e2e/conversions/change-conversion-date.cy.ts
@@ -160,7 +160,7 @@ describe("Change the conversion date tests", () => {
             .selectReasonWithDetails("Legal Documents", "Legal documents need to be finalised before the conversion.")
             .selectReasonWithDetails("Correcting an error", "The previous date was incorrect.")
             .selectReasonWithDetails("Voluntary deferral", "Being voluntarily deferred for strategic reasons.")
-            .selectReasonWithDetails("Advisory board conditions not met", "Advisory board conditions not yet met.")
+            .selectReasonWithDetails("Decision conditions not met", "Decision conditions not yet met.")
             .saveAndContinue();
 
         Logger.log("Confirmation page shows the new conversion date");