+++ CHANGELOG.md
@@ -18,6 +18,9 @@ Security in case of vulnerabilities.
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2026-03-24.1419...main) for everything awaiting release
 
+### Changed
+- remove advisory board references in change decision reasons
+
 ---
 
 ## [1.41.1](https://github.com/DFE-Digital/complete-conversions-transfers-changes/releases/tag/production-2026-03-24.1419) - 2026-03-24